+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 0.2 - 2016-07-xx
+- Split up code into several files
+- Allow user to specify start and end value for reset limit
+
 ## 0.1.1 - 2016-06-12
 
 - Allow user to specify the reset limit